+++ CHANGELOG.md
@@ -4,6 +4,13 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/).
 
 ### Unreleased
 
+### [3.1.1] - 2026-05-14
+
+- fix: strengthen control character rejection
+- fix: prevent prototype pollution during hydration
+- test: added additonal test cases
+- doc(LICENSE): added file
+
 ### [3.1.0] - 2026-05-14
 
 - feat: new RFC-5322 recursive-descent parser, O(1) performance, dependency free
@@ -36,3 +43,4 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/).
 [addr2822]: https://github.com/haraka/node-address-rfc2822
 [3.0.0]: https://github.com/haraka/email-address/releases/tag/v3.0.0
 [3.1.0]: https://github.com/haraka/email-address/releases/tag/v3.1.0
+[3.1.1]: https://github.com/haraka/email-address/releases/tag/v3.1.1